+++ app/services/forest_liana/schema_adapter.rb
@@ -227,6 +227,8 @@ module ForestLiana
         'String'
       when :citext
         'String'
+      when :uuid
+        'String'
       when :boolean
         'Boolean'
       end