+++ lib/forest_liana/collection.rb
@@ -21,7 +21,8 @@ module ForestLiana::Collection
     def field(name, opts, &block)
       model.fields << opts.merge({ field: name, :'is-searchable' => false })
 
-      if serializer_name
+      if serializer_name && ForestLiana::UserSpace.const_defined?(
+          serializer_name)
         ForestLiana::UserSpace.const_get(serializer_name).class_eval do
           attribute(name, &block)
         end
@@ -35,7 +36,8 @@ module ForestLiana::Collection
         type: ['String']
       })
 
-      if serializer_name
+      if serializer_name && ForestLiana::UserSpace.const_defined?(
+          serializer_name)
         ForestLiana::UserSpace.const_get(serializer_name).class_eval do
           has_many(name, name: name)
         end