+++ app/services/forest_liana/belongs_to_updater.rb
@@ -9,13 +9,8 @@ module ForestLiana
 
     def perform
       @record = @resource.find(@params[:id])
-
-      if @data && @data[:id]
-        new_value = @association.klass.find(@data[:id]) if @data && @data[:id]
-        @record.send("#{@association.name}=", new_value)
-      else
-        @record.send("#{@association.foreign_key}=", nil)
-      end
+      new_value = @association.klass.find(@data[:id]) if @data && @data[:id]
+      @record.send("#{@association.name}=", new_value)
 
       @record.save()
     end