+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Filters - Users want the OR filter operator with their conditions (restricted to simple conditions).
 
 ## [1.3.40] - 2016-09-26
 ### Added