+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 ### Added
 - Filters - Users want the OR filter operator with their conditions (restricted to simple conditions).
 
-## [1.3.40] - 2016-09-26
+### Fixed
+- Record Update - Fix the potential dissociations on record update.
+
+## RELEASE 1.3.40 - 2016-09-26
 ### Added
 - Filters - Users want to have "From now" and "Today" operators.