+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## RELEASE 1.3.42 - 2016-10-03
 ### Fixed
+- Smart field - Ensure the serializer exists before creating the smart field.
 - Record Create - Fix empty relationships on record creation.
 
 ## RELEASE 1.3.41 - 2016-09-26